# Conflicts:
#	CMakeLists.txt
#	Depends-Android.cmake
#	res/about/version.gmi
#	res/lang/cs.bin
#	res/lang/eo.bin
#	res/lang/es.bin
#	res/lang/fi.bin
#	res/lang/gl.bin
#	res/lang/ia.bin
#	res/lang/ie.bin
#	res/lang/ru.bin
#	res/lang/sr.bin
#	res/lang/tok.bin
#	res/lang/tr.bin
#	res/lang/uk.bin
#	res/lang/zh_Hant.bin
#	src/app.c
#	src/main.c
#	src/ui/documentwidget.c
#	src/ui/labelwidget.c
#	src/ui/sidebarwidget.c
#	src/ui/util.c

+++ src/visited.c
@@ -105,7 +105,7 @@ void load_Visited(iVisited *d, const char *dirPath) {
             char *endp = NULL;
             const unsigned long long ts = strtoull(line.start, &endp, 10);
             if (ts == 0) break;
-            const uint32_t flags = strtoul(skipSpace_CStr(endp), &endp, 16);
+            const uint32_t flags = (uint32_t) strtoul(skipSpace_CStr(endp), &endp, 16);
             const char *urlStart = skipSpace_CStr(endp);
             iVisitedUrl item;
             item.when.ts = (struct timespec){ .tv_sec = ts };
